Add files via upload
here is a Group presentation that was done in a team environment where i have utilized SQL Queries in GCP to find out why most of Americans are suffering from Obesity. The data was collected from a national institute of Health. Data was cleaned in using Open Refine and then imported in GCP SQL.